Create modules for chungo, garfanzo, and fatkhiyev
